refactor(schema): isolate v1 contracts
